src/requirements.txt
fastapi==0.110.2

uvicorn[standard]==0.29.0

python-multipart==0.0.9

python-dotenv==1.0.1

pydantic-settings==2.2.1

aiofiles==23.2.1

langchain==0.1.20

PyMuPDF==1.24.3
